Embed unicode list instead downloading it
Also switched the list to emoji-test.txt which contains all emojis
which the previous list didn't for some reason.

As a consequence "range" support is removed from the parse code. Doesn't
really affect end-users though, but only an internal detail.

Since the emojis are embedded, they need to be downloaded before
building. See the 'build' job in the Makefile for details on that is
done.

Also updated the example printer and moved it to an example folder